storage: move test-only code out of live code
